refactor: enhance VideoBlock schema and introduce new video source enum
- Added a type guard for media video sources in VideoBlock to improve validation.
- Refactored the VideoBlockObjectFit enum to use Prisma definitions for consistency.
- Introduced a new VideoBlockSource enum to standardize video source handling.
- Removed redundant video source enum from mediaVideo schema to streamline imports.
