+++ apis/api-journeys-modern/src/schema/block/video/enums/videoSource.ts
@@ -1,6 +1,6 @@
 import { VideoBlockSource as PrismaVideoSource } from '.prisma/api-journeys-modern-client'
 
-import { builder } from '../../builder'
+import { builder } from '../../../builder'
 
 export const VideoBlockSource = builder.enumType(PrismaVideoSource, {
   name: 'VideoBlockSource'
