--- apis/api-journeys-modern/src/schema/mediaVideo/enums/index.ts
@@ -1,1 +1,0 @@
-import './videoSource'
