+++ apis/api-journeys-modern/src/schema/mediaVideo/index.ts
@@ -1,2 +1,1 @@
-import './enums'
 import './mediaVideo'
